cambio de Integer a Long para idProducto

+++ sportWear/src/main/java/com/ecommerce/sportWear/repository/ProductoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.ecommerce.sportWear.model.Producto;
 
-public interface ProductoRepository extends MongoRepository<Producto, Integer> {
+public interface ProductoRepository extends MongoRepository<Producto, Long> {
 	
 }
